# api-gateway-user/docs/swagger.yaml
basePath: /
definitions:
...
  genproto.Category:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
...
  genproto.Course:
    properties:
      category:
        $ref: '#/definitions/genproto.Category'
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
...
  genproto.Pagination:
    properties:
      limit:
        type: integer
      offset:
        type: integer
    type: object
...
  genproto.UserCourse:
    properties:
      course:
        $ref: '#/definitions/genproto.Course'
      end_date:
        type: string
      id:
        type: string
      start_date:
        type: string
      status:
        type: string
      user:
        $ref: '#/definitions/genproto.UserGetRes'
    type: object
...
  genproto.UserGetRes:
    properties:
      address:
        type: string
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      id:
        type: string
      is_confirmed:
        type: boolean
      last_name:
        type: string
      occupation:
        type: string
      phone_number:
        type: string
      photo_url:
        type: string
    type: object
  genproto.UserLesson:
    properties:
      LessonId:
        type: string
      UserCourse:
        $ref: '#/definitions/genproto.UserCourse'
      id:
        type: string
      status:
        description: 'Enum type: progress_status_type'
        type: string
    type: object
  genproto.UserLessonCreateReq:
    properties:
      UserCourseId:
        type: string
      lesson_id:
        type: string
      status:
        type: string
    type: object
  genproto.UserLessonListsRes:
    properties:
      TotalCount:
        type: integer
      pagination:
        $ref: '#/definitions/genproto.Pagination'
      user_lessons:
        items:
          $ref: '#/definitions/genproto.UserLesson'
        type: array
    type: object
  genproto.UserLessonUpdateReq:
    properties:
      body:
        $ref: '#/definitions/genproto.UserUptbody'
      id:
        type: string
    type: object
...
  genproto.UserUptbody:
    properties:
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: API for Instant Delivery resources
  title: Finance Tracker API Documentation
  version: "1.0"
paths:
...
  /user-lessons/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a user lesson by its ID
      parameters:
      - description: User Lesson ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User Lesson details
          schema:
            $ref: '#/definitions/genproto.UserLesson'
        "400":
          description: Invalid request
          schema:
            type: string
        "404":
          description: User Lesson not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get User Lesson
      tags:
      - UserLesson
  /user-lessons/create:
    post:
      consumes:
      - application/json
      description: Create a new user lesson
      parameters:
      - description: User Lesson data
        in: body
        name: user_lesson
        required: true
        schema:
          $ref: '#/definitions/genproto.UserLessonCreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: User Lesson created successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create User Lesson
      tags:
      - UserLesson
  /user-lessons/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user lesson by its ID
      parameters:
      - description: User Lesson ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User Lesson deleted successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete User Lesson
      tags:
      - UserLesson
  /user-lessons/list:
    get:
      consumes:
      - application/json
      description: Retrieve a list of user lessons with optional filters
      parameters:
      - description: User Course ID
        in: query
        name: user_course_id
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of user lessons
          schema:
            $ref: '#/definitions/genproto.UserLessonListsRes'
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List User Lessons
      tags:
      - UserLesson
  /user-lessons/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a user lesson
      parameters:
      - description: User Lesson ID
        in: path
        name: id
        required: true
        type: string
      - description: User Lesson update data
        in: body
        name: user_lesson
        required: true
        schema:
          $ref: '#/definitions/genproto.UserLessonUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: User Lesson updated successfully
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update User Lesson
      tags:
      - UserLesson
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"